helios_aeo/requirements.txt
openai>=1.3.0
requests>=2.32.0
beautifulsoup4>=4.12.0
python-dotenv>=1.0.1
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
pydantic>=2.5.0
gunicorn>=21.2.0
